Added Pumpkins and Melons and Expanded Cactus Functionality
For Pumpkins and Melons, they now grow an additional fruit when the faction upgrade is applied. Extra fruit generation mimics vanilla requirements, except, when activated, the upgrade will always grow an additional fruit if there is space around it and the vanilla conditions are met. The extra fruit is also grown fast enough to be compatible with redstone auto farms.

For cactus, if there are neighboring blocks around where the cactus naturally grew, the factions upgrade will now wait until the cactus is broken and replant an additional cactus in its place, which will also break, leading to an extra cactus dropped. In combination with the height increase, this update should now provide an extra cactus in all circumstances.