From Dak, Untested with Commit Comment
Make material sets lazy-initialized and clean up function parameters

Material sets, such as ManualCrops, are now lazy-initialized, meaning the set is now class-scoped,
and is not created until it is first needed. This means the set will only ever be created once,
rather than being created each time the function is called, which is redundant because the function
will always use the same set.

Function parameters were cleaned up in cases where a function used a variable that was needlessly
class-scoped and could rather be passed into that function by the parent function calling it.
Example: The calls to growManual, growFruit, and growAutoAgeable in growCrop() are now passed in
the block data rather than grabbing it from the class variables.